refactor: Validator 클래스 util 패키지로 이동
