+++ src/main/java/racingcar/controller/GameController.java
@@ -1,7 +1,7 @@
 package racingcar.controller;
 
 import racingcar.model.Car;
-import racingcar.model.Validator;
+import racingcar.util.Validator;
 import racingcar.service.CarService;
 import racingcar.service.GameService;
 import racingcar.view.InputView;
